# src/main/java/com/example/demo/dto/AdministratorUpdateRequest.java
package com.example.demo.dto;

public class AdministratorUpdateRequest {
    private String name;
    private String mailAddress;
    private String password;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getMailAddress() {
        return mailAddress;
    }

    public void setMailAddress(String mailAddress) {
        this.mailAddress = mailAddress;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
}